chromium/third_party/blink/web_tests/typedcssom/computedstyle/transform.html

<!DOCTYPE html>
<html>
<head>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
</head>
<style>
  /* regression test for failing to unzoom transform coefficients */
  body { zoom: 1.5; }
</style>
<body>
<div id='testElement'></div>

<script>

var testElement = document.getElementById('testElement');
var computedStyleMap = testElement.computedStyleMap();

test(function() {
  var result = computedStyleMap.get('transform');
  assert_equals(result.toString(), 'none');
}, 'transform is none by default');

test(function() {
  testElement.style.transform = 'translate(100px, 30%)';
  var result = computedStyleMap.get('transform');
  assert_equals(result.toString(), 'translate(100px, 30%)');
}, 'transform preserves a translate value');

test(function() {
  testElement.style.transform = 'translate3d(100px, 30%, 40px)';
  var result = computedStyleMap.get('transform');
  assert_equals(result.toString(), 'translate3d(100px, 30%, 40px)');
}, 'transform preserves a translate3d value');

test(function() {
  testElement.style.transform = 'rotate(90deg)';
  var result = computedStyleMap.get('transform');
  assert_equals(result.toString(), 'rotate(90deg)');
}, 'transform preserves a rotate value');

test(function() {
  testElement.style.transform = 'rotate3d(0, 0, 1, 90deg)';
  var result = computedStyleMap.get('transform');
  assert_equals(result.toString(), 'rotate3d(0, 0, 1, 90deg)');
}, 'transform preserves a rotate3d value');

test(function() {
  testElement.style.transform = 'scale(1, 2)';
  var result = computedStyleMap.get('transform');
  assert_equals(result.toString(), 'scale(1, 2)');
}, 'transform preserves a scale value');

test(function() {
  testElement.style.transform = 'scale3d(1, 2, 3)';
  var result = computedStyleMap.get('transform');
  assert_equals(result.toString(), 'scale3d(1, 2, 3)');
}, 'transform preserves a scale3d value');

test(function() {
  testElement.style.transform = 'perspective(100px)';
  var result = computedStyleMap.get('transform');
  assert_equals(result.toString(), 'perspective(100px)');
}, 'transform preserves a perspective value');

test(function() {
  testElement.style.transform = 'skew(10deg, 10deg)';
  var result = computedStyleMap.get('transform');
  assert_equals(result.toString(), 'skew(10deg, 10deg)');
}, 'transform preserves a skew value');

test(function() {
  testElement.style.transform = 'matrix(1, 0, 0, 2, 3, 4)';
  var result = computedStyleMap.get('transform');
  assert_equals(result.toString(), 'matrix(1, 0, 0, 2, 3, 4)');
}, 'transform preserves a matrix value');

test(function() {
  testElement.style.transform = 'matrix3d(1, 0, 0, 0, 0, 2, 0, 0, 0, 0, 3, 0, 4, 5, 6, 1)';
  var result = computedStyleMap.get('transform');
  assert_equals(result.toString(), 'matrix3d(1, 0, 0, 0, 0, 2, 0, 0, 0, 0, 3, 0, 4, 5, 6, 1)');
}, 'transform preserves a matrix3d value');

test(function() {
  testElement.style.transform = 'matrix3d(1, 0, 0, 0.01, 0, 2, 0, 0.02, 0, 0, 3, 0.03, 4, 5, 6, 1)';
  var result = computedStyleMap.get('transform');
  assert_equals(result.toString(), 'matrix3d(1, 0, 0, 0.01, 0, 2, 0, 0.02, 0, 0, 3, 0.03, 4, 5, 6, 1)');
}, 'transform preserves a matrix3d value with perspective');

</script>
</body>
</html>
